bot/pyproject.toml
[tool.poetry]
name = "cogdl-bot"
version = "0.1.0"
description = ""
authors = ["Parker Freestone <parkerfreestone@gmail.com>"]
readme = "README.md"
packages = [
    { include = "cogs", from = "src" },
    { include = "ui", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.12"
aiohttp = "^3.10.10"
python-dotenv = "^1.0.1"
discord-py = "^2.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
